refactor(core): drop aiprovider re-export facade, import the leaf directly
- migrate remaining internal/core/aiprovider importers to internal/aiprovider
- delete the transitional facade left by #508
